refactor path to storage-wrapper
