+++ app/core/SDKConnect/AndroidSDK/addDappConnection.test.ts
@@ -5,7 +5,7 @@ import addDappConnection from './addDappConnection';
 jest.mock('../Connection');
 jest.mock('../SDKConnect');
 jest.mock('../utils/DevLogger');
-jest.mock('../../../store/async-storage-wrapper', () => ({
+jest.mock('../../../store/storage-wrapper', () => ({
   setItem: jest.fn().mockResolvedValue(''),
 }));
 jest.mock('../../../core/AppConstants');
