+++ app/core/SDKConnect/ConnectionManagement/approveHost.test.ts
@@ -8,7 +8,7 @@ jest.mock('../../../core/AppConstants');
 jest.mock('../SDKConnect');
 jest.mock('../SDKConnectConstants');
 jest.mock('../utils/DevLogger');
-jest.mock('../../../store/async-storage-wrapper', () => ({
+jest.mock('../../../store/storage-wrapper', () => ({
   setItem: jest.fn().mockResolvedValue([]),
   getItem: jest.fn().mockResolvedValue(JSON.stringify({})),
 }));
