+++ app/core/SDKConnect/ConnectionManagement/connectToChannel.test.ts
@@ -3,7 +3,7 @@ import { DEFAULT_SESSION_TIMEOUT_MS } from '../SDKConnectConstants';
 import { SDKConnect } from './../SDKConnect';
 import connectToChannel from './connectToChannel';
 
-jest.mock('../../../store/async-storage-wrapper', () => ({
+jest.mock('../../../store/storage-wrapper', () => ({
   setItem: jest.fn().mockResolvedValue(''),
   getItem: jest.fn().mockResolvedValue(''),
 }));
