+++ app/core/SDKConnect/ConnectionManagement/removeChannel.test.ts
@@ -5,7 +5,7 @@ import removeChannel from './removeChannel';
 jest.mock('../../../core/AppConstants');
 jest.mock('../SDKConnect');
 jest.mock('../utils/DevLogger');
-jest.mock('../../../store/async-storage-wrapper', () => ({
+jest.mock('../../../store/storage-wrapper', () => ({
   setItem: jest.fn().mockResolvedValue([]),
   getItem: jest.fn().mockResolvedValue(JSON.stringify({})),
 }));
