+++ app/util/Logger/index.ts
@@ -3,7 +3,7 @@ import {
   captureException,
   withScope,
 } from '@sentry/react-native';
-import StorageWrapper from '../../store/async-storage-wrapper';
+import StorageWrapper from '../../store/storage-wrapper';
 import { METRICS_OPT_IN, AGREED, DEBUG } from '../../constants/storage';
 
 interface ExtraInfo {
